Delete unused tests.
